Refactor type imports to use moved interfaces from the types directory
