+++ src/app/(faculties)/fcyt/[year]/page.tsx
@@ -4,7 +4,7 @@ import { isValidYear } from "@/utils/validation";
 import { getFcytExamsData } from "@/utils/services";
 import { redirect } from "next/navigation";
 
-import type { Exam, OptionType } from "@/utils/interfaces";
+import type { Exam, OptionType } from "@/types/interfaces";
 
 const examsData: Exam[] = await getFcytExamsData();
