+++ src/utils/services.ts
@@ -1,4 +1,4 @@
-import type { Exam } from "./interfaces";
+import type { Exam } from "../types/interfaces";
 
 export const getFcytExamsData = async (): Promise<Exam[]> => {
   const response = await fetch(
