+++ src/utils/validation.ts
@@ -1,4 +1,4 @@
-import type { OptionType } from "./interfaces";
+import type { OptionType } from "../types/interfaces";
 
 export const isValidYear = (year: number, yearsOptions: OptionType[]) =>
   yearsOptions.some((option) => option.value === year);
